refactor: move Docker package to internal
